Add more filtering tests
